requirements.txt
aiogram
anthropic
python-dotenv
gspread-asyncio
google-auth
tenacity